fix(deepseek-v4): read FP32 attn_sink as float, not bf16 — restores 7 hard-zeroed heads (#341)
Ported onto the active integration line (main). The checkpoint ships
layers.N.attn.attn_sink as F32 [num_q_heads]; the DS4F sink-consuming kernels
declared/indexed it as const __nv_bfloat16*. A 2-byte stride over the 4-byte fp32
buffer makes even head 2k read the low-mantissa half of fp32 element k; for
{6,10,12,20,28,34,48} that misread decodes large-positive, the sink dominates the
softmax denominator and exp(m-sink)->0 annihilates the value accumulator -> exact
zero output. Reading the banked device sink bytes through the kernel's own
arithmetic predicts those 7 heads deterministically. vLLM/FP32-ref keep all heads.

Loader: new weight_loader/deepseek_v4/attn_sink.rs normalizes attn_sink to the
canonical device FP32 buffer (F32 pass-through / BF16 widen once / else fail /
missing NULL) + bf16_bytes_to_f32_bytes helper + regression tests. Kernels: the 3
sink consumers on this line (inferspark_prefill_512, prefill_attn_compressed,
mla_paged_decode_fp8) index const float* (drop __bfloat162float/(float) cast). ABI
unchanged (Rust passes a raw DevicePtr). No 4th kernel (mla_paged_decode_nvfp4 not
on this line). Tests: attn_sink_dtype_tests 3/3.

+++ kernels/gb10/deepseek-v4-flash/nvfp4/inferspark_prefill_512.cu
@@ -21,7 +21,7 @@ extern "C" __global__ void inferspark_prefill_512(
     const float inv_sqrt_d,
     const unsigned int causal,
     const unsigned int sliding_window,  // Always 0 for full-attention 512-hd layers; param added for signature consistency
-    const __nv_bfloat16* __restrict__ sinks  // [num_q_heads] per-head sink logit (denominator only); nullptr = no sink
+    const float* __restrict__ sinks  // [num_q_heads] per-head sink logit (denominator only); nullptr = no sink. FP32: checkpoint-native dtype (reading as bf16 hard-zeroed 7 heads)
 ) {
     const unsigned int q_head = blockIdx.x;
     const unsigned int q_block = blockIdx.y;
@@ -118,7 +118,7 @@ extern "C" __global__ void inferspark_prefill_512(
     // attention (non-CSA) prefill layers were missing it, so prefill diverged
     // from the sink-applying decode path and corrupted the prompt's hidden/KV.
     if (sinks != nullptr) {
-        float sg = __bfloat162float(sinks[q_head]);
+        float sg = sinks[q_head];
         float m_new = fmaxf(m, sg);
         float exp_old = __expf(m - m_new);
         float exp_sink = __expf(sg - m_new);
